registration dialog & alert warning style
