+++ frontend/app/(planner)/[planner]/@registration/(.)register/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import Modal from "@/components/modal";
+import RegisterForm from "../../register/register-form";
+
+export default function Page() {
+  return (
+    <Modal>
+      <RegisterForm modal={true} />
+    </Modal>
+  );
+}
